requirements.txt
botbuilder-core
aiohttp
langchain
langchain-chroma
langchain-openai
langchain-google-vertexai
langchain-google-genai
langchain-community
slack-bolt
python-dotenv
chromadb
tiktoken
nest_asyncio

